refactor: extract client monitoring bootstrap
