+++ src/main.ts
@@ -12,7 +12,6 @@ import {
   hasDistrictSessionHandoff,
   refreshPublicAuthFromSession,
 } from "./services/publicAuthBootstrap";
-import { allowsAnalytics, allowsDiagnostics, readCookieConsent } from "./services/cookieConsentService";
 import { TimeoutError, withTimeout } from "./services/asyncUtils";
 import {
   captureInitialPerfMetrics,
@@ -27,6 +26,7 @@ import {
   isAdminBootstrapRoute,
   isPublicBootstrapRoute,
 } from "./bootstrap/routeBootstrap";
+import { createClientMonitoring } from "./bootstrap/clientMonitoring";
 
 const redirectCanonicalHost = () => {
   if (typeof window === "undefined") return false;
@@ -109,103 +109,7 @@ const initializePublicAuth = async () => {
   }
 };
 
-let appMounted = false;
-
-const initializeSentry = async (app: ReturnType<typeof createApp>) => {
-  if (!import.meta.env.VITE_SENTRY_DSN?.trim() || !allowsDiagnostics(readCookieConsent())) {
-    return;
-  }
-  try {
-    const { initializeSentry: initializeSentryMonitoring } = await import("./services/sentry");
-    await initializeSentryMonitoring(app, router, appMounted);
-  } catch (error) {
-    // Diagnostics must never break login or core flows.
-    console.warn("[sentry] initialization failed; continuing without diagnostics.", error);
-  }
-};
-
-const initializePostHog = async () => {  
-  if (!import.meta.env.VITE_POSTHOG_PROJECT_TOKEN?.trim() || !allowsAnalytics(readCookieConsent())) {
-    return;
-  }
-  try {
-    const { initPostHog } = await loadPostHogService();
-    await initPostHog();
-  } catch (error) {
-    // Analytics must never break login or core flows.
-    console.warn("[posthog] initialization failed; continuing without analytics.", error);
-  }
-};
-
-const initializeClientDiagnostics = async () => {
-  if (!allowsDiagnostics(readCookieConsent())) {
-    return;
-  }
-  try {
-    const { installClientDiagnostics } = await import("./services/clientDiagnostics");
-    installClientDiagnostics();
-  } catch (error) {
-    // Diagnostics must never break login or core flows.
-    console.warn("[diagnostics] initialization failed; continuing without diagnostics.", error);
-  }
-};
-
-let posthogServicePromise: Promise<typeof import("./services/posthogService")> | null = null;
-
-const loadPostHogService = () => {
-  if (!posthogServicePromise) {
-    posthogServicePromise = import("./services/posthogService").catch((error) => {
-      posthogServicePromise = null;
-      throw error;
-    });
-  }
-  return posthogServicePromise;
-};
-
-const getPostHogExceptionCapture = async () => {
-  if (!allowsAnalytics(readCookieConsent())) {
-    return () => undefined;
-  }
-  return loadPostHogService()
-    .then((module) => module.capturePostHogException)
-    .catch(() => () => undefined);
-};
-
-const capturePostHogExceptionSafely = (error: unknown) => {
-  void getPostHogExceptionCapture().then((capture) => capture(error));
-};
-
-const bindConsentDrivenMonitoring = (app: ReturnType<typeof createApp>) => {
-  const maybeEnableDiagnostics = () => {
-    if (!allowsDiagnostics(readCookieConsent())) {
-      return;
-    }
-    void initializeSentry(app);
-    void initializeClientDiagnostics();
-  };
-
-  const maybeEnableAnalytics = () => {
-    if (allowsAnalytics(readCookieConsent())) {
-      void loadPostHogService()
-        .then(({ syncPostHogConsent }) => {
-          syncPostHogConsent();
-          void initializePostHog();
-        })
-        .catch((error) => {
-          console.warn("[posthog] consent sync failed; continuing without analytics.", error);
-        });
-      return;
-    }
-    if (posthogServicePromise) {
-      void posthogServicePromise
-        .then(({ syncPostHogConsent }) => syncPostHogConsent())
-        .catch(() => undefined);
-    }
-  };
-
-  window.addEventListener("itemtraxx:cookie-consent", maybeEnableDiagnostics);
-  window.addEventListener("itemtraxx:cookie-consent", maybeEnableAnalytics);
-};
+const clientMonitoring = createClientMonitoring(router);
 
 const mountApp = async () => {
   markRouteNavigationStart();
@@ -225,25 +129,16 @@ const mountApp = async () => {
   const app = createApp(App);
   const existingErrorHandler = app.config.errorHandler;
   app.config.errorHandler = (error, instance, info) => {
-    capturePostHogExceptionSafely(error);
+    clientMonitoring.captureException(error);
     if (existingErrorHandler) {
       existingErrorHandler(error, instance, info);
     }
   };
   app.use(router);
   await router.isReady();
-  await initializeSentry(app);
+  await clientMonitoring.initializeBeforeMount(app);
   app.mount("#app");
-  appMounted = true;
-  void initializeClientDiagnostics().catch(() => undefined);
-  void import("./services/globalErrorHandling")
-    .then(({ installGlobalErrorHandling }) => installGlobalErrorHandling(app))
-    .catch(() => undefined);
-  void import("./services/appErrorRecovery")
-    .then(({ installAppErrorRecovery }) => installAppErrorRecovery(router))
-    .catch(() => undefined);
-  void initializePostHog();
-  bindConsentDrivenMonitoring(app);
+  clientMonitoring.initializeAfterMount(app);
   captureInitialPerfMetrics();
   if (import.meta.env.VITE_E2E_TEST_UTILS === "true") {
     const { attachE2EControls } = await import("./e2e/testControls");
